Timer and Scene Nav fixes
Saved player data to SceneNav to be displayed at the end of each day, displayed on success screen. Player is no longer able to move to the cook screen unless they have accepted an order from a customer. This prevents the player trying to preemptively make food until they receive an order